requirements.txt
requests
google-api-python-client
fuzzywuzzy
python-Levenshtein
google-cloud-language